pyproject.toml
[tool.poetry]
name = "protoforge"
version = "0.1.0"
description = "Elite AI-powered CLI for virtual hardware prototyping."
authors = ["ProtoForge Contributors"]
license = "MIT"
readme = "README.md"
packages = [{include = "protoforge"}]

[tool.poetry.dependencies]
python = "^3.12"
typer = {extras = ["all"], version = "^0.12.3"}
rich = "^13.7.1"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
litellm = "^1.37.0"
sqlmodel = "^0.0.18"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
cadquery = {version = "^2.4.0", optional = true}
build123d = {version = "^0.4.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.4.4"
mypy = "^1.10.0"
pre-commit = "^3.7.1"

[tool.poetry.scripts]
proto = "protoforge.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
